update be and UI of Shift Management

+++ HRM-be/src/main/java/com/example/hrm/dto/request/ShiftRuleRequest.java
@@ -11,5 +11,5 @@ public class ShiftRuleRequest {
     private String description;
     private Integer maxHoursPerDay;
     private Boolean allowOvertime;
-    private Double nightShiftMultiplier;
+    private Float nightShiftMultiplier;
 }
